requirements.txt
pillow<7.0
cython
tensorboardX
numpy
umsgpack
torchvision==0.2.2
tqdm
scikit-image
pycocotools
cityscapesscripts
git+git://github.com/waspinator/pycococreator.git@0.2.0
git+https://github.com/mapillary/inplace_abn.git
